add sites bundle
